# cdc_erp_updated/cdc_erpp1/cdc_erpp/Document-Type-Designs/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDC ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /release-records:
    get:
      operationId: listReleaseRecords
      tags: [release-records]
      summary: List release records
      parameters:
        - name: childId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of release records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReleaseRecord"
    post:
      operationId: createReleaseRecord
      tags: [release-records]
      summary: Create a release record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseRecordBody"
      responses:
        "201":
          description: Created release record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseRecord"
  /release-records/{id}:
    get:
      operationId: getReleaseRecord
      tags: [release-records]
      summary: Get a release record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseRecord"

  # Follow-up Logs
...
  /risk-assessments:
    get:
      operationId: listRiskAssessments
      tags: [risk-assessments]
      summary: List risk assessments
      parameters:
        - name: childId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of risk assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskAssessment"
    post:
      operationId: createRiskAssessment
      tags: [risk-assessments]
      summary: Create a risk assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRiskAssessmentBody"
      responses:
        "201":
          description: Created risk assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"
  /risk-assessments/{id}:
    get:
      operationId: getRiskAssessment
      tags: [risk-assessments]
      summary: Get a risk assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"
    put:
      operationId: updateRiskAssessment
      tags: [risk-assessments]
      summary: Update a risk assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRiskAssessmentBody"
      responses:
        "200":
          description: Updated risk assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"
...
components:
  schemas:
...
    ReleaseRecord:
      type: object
      properties:
        id:
          type: integer
        releaseId:
          type: string
        childId:
          type: integer
        childName:
          type: string
        releaseDate:
          type: string
          format: date
        releaseType:
          type: string
        handedOverTo:
          type: string
        authorityApproval:
          type: boolean
        remarks:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - releaseId
        - childId
        - releaseDate
        - releaseType
        - createdAt
    CreateReleaseRecordBody:
      type: object
      properties:
        childId:
          type: integer
        releaseDate:
          type: string
          format: date
        releaseType:
          type: string
        handedOverTo:
          type: string
        authorityApproval:
          type: boolean
        remarks:
          type: string
      required:
        - childId
        - releaseDate
        - releaseType
...
    RiskAssessment:
      type: object
      properties:
        id:
          type: integer
        riskId:
          type: string
        childId:
          type: integer
        childName:
          type: string
        assessmentDate:
          type: string
          format: date
        assessedBy:
          type: string
        abuseRisk:
          type: string
          enum: [Low, Medium, High]
        traffickingRisk:
          type: string
          enum: [Low, Medium, High]
        reoffendingRisk:
          type: string
          enum: [Low, Medium, High]
        selfHarmRisk:
          type: string
          enum: [Low, Medium, High]
        overallRiskLevel:
          type: string
          enum: [Low, Medium, High]
        immediateActionRequired:
          type: boolean
        protectionMeasures:
          type: string
        status:
          type: string
          enum: [Draft, Reviewed, Approved]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - riskId
        - childId
        - assessmentDate
        - overallRiskLevel
        - createdAt
    CreateRiskAssessmentBody:
      type: object
      properties:
        childId:
          type: integer
        assessmentDate:
          type: string
          format: date
        assessedBy:
          type: string
        abuseRisk:
          type: string
          enum: [Low, Medium, High]
        traffickingRisk:
          type: string
          enum: [Low, Medium, High]
        reoffendingRisk:
          type: string
          enum: [Low, Medium, High]
        selfHarmRisk:
          type: string
          enum: [Low, Medium, High]
        overallRiskLevel:
          type: string
          enum: [Low, Medium, High]
        immediateActionRequired:
          type: boolean
        protectionMeasures:
          type: string
        status:
          type: string
          enum: [Draft, Reviewed, Approved]
      required:
        - childId
        - assessmentDate
        - overallRiskLevel